Rename some files for autodoc
