+++ mepo.d/command/stash/list/list.py
