+++ mepo.d/command/stash/pop/pop.py
