+++ mepo.d/command/stash/show/show.py
