+++ mepo.d/command/stash/stash.py
@@ -2,18 +2,18 @@ import subprocess as sp
 
 from state.state import MepoState
 
-from command.stash.stlist  import stlist
-from command.stash.stpop   import stpop
-from command.stash.stapply import stapply
-from command.stash.stpush  import stpush
-from command.stash.stshow  import stshow
+from command.stash.list  import list
+from command.stash.pop   import pop
+from command.stash.apply import apply
+from command.stash.push  import push
+from command.stash.show  import show
 
 def run(args):
     d = {
-        'list':  stlist,
-        'pop':   stpop,
-        'apply': stapply,
-        'push':  stpush,
-        'show':  stshow,
+        'list':  list,
+        'pop':   pop,
+        'apply': apply,
+        'push':  push,
+        'show':  show,
     }
     d[args.mepo_stash_cmd].run(args)
