+++ mepo.d/command/tag/tag.py
@@ -2,14 +2,14 @@ import subprocess as sp
 
 from state.state import MepoState
 
-from command.tag.tglist import tglist
+from command.tag.list   import list
 from command.tag.create import create
 from command.tag.delete import delete
-from command.tag.push import push
+from command.tag.push   import push
 
 def run(args):
     d = {
-        'list': tglist,
+        'list': list,
         'create': create,
         'delete': delete,
         'push': push
